trying to add more game dialogue and events
nature fairy comes to the rescue of the hapless soul about to be destroyed by the evil sorceror. The sorceror gets blasted by a thunderbolt and a path is revealed to escape the demon castle.

Still not quite qorking. There seem to be bugs with the event manager. The fairy is beginning her dialogue immediately on arrival to demon castle.